cleaned up components directory
