+++ src/components/addItem/NewProductInfo.js
