+++ src/components/common/index.js
@@ -2,15 +2,7 @@ import FormButton from './FormButton';
 import FormInput from './FormInput';
 import List from './List';
 import LoadingComponent from './LoadingComponent';
-import TestItemImageUpload from './TestItemImageUpload';
 import Button from './Button';
 // notice we're building out a 'package' of reusables here and exporting them as an object of component properties.
 // to use this, simply `import {foo, bar, baz} from '<path-to-this-directory>/ReusableComponents';`
-export {
-  FormButton,
-  FormInput,
-  List,
-  LoadingComponent,
-  Button,
-  TestItemImageUpload,
-};
+export { FormButton, FormInput, List, LoadingComponent, Button };
