--- src/components/pages/Login/README.md
@@ -1,7 +1,0 @@
-## Login and Authentication w/ Okta.
-
-- The `LoginContainer.js` file includes a widget that is provided to us by Okta.
-- This widget is the key to `user authentication and management` within your application.
-- You don't have to manage ANY users because we're letting Okta do this for us :)
-- This component is heavily documented with comments but it'd be REALLY good to read through [the docs here](https://github.com/okta/okta-signin-widget) to get a good understanding of the configuration we're doing with the widget itself.
-- You'll also want to learn a bit about how to style this widget to meet your app's concerns.
