--- src/components/pages/Login/index.js
@@ -1,1 +1,0 @@
-export { default as LoginPage } from './LoginContainer';
