--- src/components/pages/ProductPage/ProductCarousel.js
@@ -1,48 +1,0 @@
-import React, { Component } from 'react';
-import { Carousel } from 'antd';
-import './productPage.css';
-
-export default class ProductCarousel extends Component {
-  constructor(props) {
-    super(props);
-    this.next = this.next.bind(this);
-    this.previous = this.previous.bind(this);
-    this.carousel = React.createRef();
-  }
-  next() {
-    this.carousel.next();
-  }
-  previous() {
-    this.carousel.prev();
-  }
-
-  render() {
-    const props = {
-      dots: true,
-      infinite: true,
-      speed: 500,
-      slidesToShow: 1,
-      slidesToScroll: 1,
-    };
-    return (
-      <div>
-        {/* <p onClick={this.previous}>previous</p> */}
-        <Carousel ref={node => (this.carousel = node)} {...props}>
-          <div>
-            <h3>1</h3>
-          </div>
-          <div>
-            <h3>2</h3>
-          </div>
-          <div>
-            <h3>3</h3>
-          </div>
-          <div>
-            <h3>4</h3>
-          </div>
-        </Carousel>
-        {/* <p onClick={this.next}>next</p> */}
-      </div>
-    );
-  }
-}
