--- src/components/pages/ProductPage/index.js
@@ -1,1 +1,0 @@
-export { default as ProductPage } from './ProductPage';
