--- src/components/pages/ProfileList/index.js
@@ -1,1 +1,0 @@
-export { default as ProfileListPage } from './ProfileListContainer';
