+++ src/components/sellerDashboard/SellerProfile.js
@@ -1,0 +1,14 @@
+import React from 'react';
+import NavBar from '../common/navBar';
+import SellerDashboard from './SellerDashboard';
+
+function SellerProfile() {
+  return (
+    <>
+      <NavBar />
+      <SellerDashboard />
+    </>
+  );
+}
+
+export default SellerProfile;
