--- src/components/sellerPages/profile/index.js
@@ -1,14 +1,0 @@
-import React from 'react';
-import NavBar from '../../common/navBar';
-import Dashboard from './dashboard';
-
-function SellerProfile() {
-  return (
-    <>
-      <NavBar />
-      <Dashboard />
-    </>
-  );
-}
-
-export default SellerProfile;
