+++ src/index.js
@@ -6,26 +6,23 @@ import {
   useHistory,
   Switch,
 } from 'react-router-dom';
-
-import { Security, LoginCallback, SecureRoute } from '@okta/okta-react';
-import 'antd/dist/antd.less';
-
-import { NotFoundPage } from './components/pages/NotFound';
-
-import { LoginPage } from './components/pages/Login';
-import { config } from './utils/oktaConfig';
-import reducer from './state/reducers';
 import { createStore, applyMiddleware, compose } from 'redux';
 import { Provider } from 'react-redux';
 import thunk from 'redux-thunk';
-import Landing from './components/pages/Landing/Landing';
+import { Security, LoginCallback, SecureRoute } from '@okta/okta-react';
+import 'antd/dist/antd.less';
 
-// Seller Imports
-import SellerProfile from './components/sellerPages/profile';
-import Inventory from './components/sellerPages/inventory';
-import CurrentInventory from './components/sellerPages/inventory/current';
-import { ProductPage } from './components/pages/ProductPage';
-import { TestItemImageUpload } from './components/common';
+import { config } from './utils/oktaConfig';
+import reducer from './state/reducers';
+
+import LoginPage from './components/LoginPage';
+import LandingPage from './components/LandingPage';
+import SellerProfile from './components/sellerDashboard/SellerProfile';
+import AddItem from './components/addItem/AddItem';
+import CurrentInventory from './components/CurrentInventory';
+import ProductPage from './components/ProductPage';
+import TestItemImageUpload from './components/TestItemImageUpload';
+import NotFoundPage from './components/NotFoundPage';
 
 const store = createStore(reducer, compose(applyMiddleware(thunk)));
 
@@ -57,7 +54,7 @@ function App() {
         <Route path="/login" component={LoginPage} />
         <Route path="/implicit/callback" component={LoginCallback} />
         {/* any of the routes you need secured should be registered as SecureRoutes */}
-        <Route exact path="/" component={Landing} />
+        <Route exact path="/" component={LandingPage} />
         <SecureRoute exact path="/myprofile" component={SellerProfile} />
         <SecureRoute
           exact
@@ -67,7 +64,7 @@ function App() {
         <SecureRoute
           exact
           path="/myprofile/inventory/additem"
-          component={Inventory}
+          component={AddItem}
         />
         <SecureRoute
           exact
