+++ src/styles/inventory.css
